Fix audio being cut off in mic.html worker example.
The MDN docs on createScriptProcessor recommends developers allow the browser to choose
the best bufferSize.  It seems large buffer sizes reduce the load
but also reduce quality.  Prior to this commit the maximum possible buffer size was chosen
and that seemed to result in the last part of the audio being clipped off in the recording.

Also worth noting is that apparently the ScriptProcessorNode was deprecated in 2014
but the replacement, Audio Workers, has yet to be implemented in any browsers.